refactor(codemode): split runtime into focused interpreter modules
